+++ app.json
@@ -7,6 +7,7 @@
     "icon": "./assets/icon.png",
     "userInterfaceStyle": "light",
     "newArchEnabled": true,
+    "":"",
     "splash": {
       "image": "./assets/splash-icon.png",
       "resizeMode": "contain",
