+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "curso-react-native-ocupacion",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -16,7 +16,10 @@
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.6",
-    "react-native-web": "~0.19.13"
+    "react-native-web": "~0.19.13",
+    "expo-router": "~4.0.17",
+    "react-native-screens": "~4.4.0",
+    "expo-linking": "~7.0.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
