+++ src/app/_layout.js
@@ -1,0 +1,8 @@
+import { Slot } from "expo-router";
+import { View } from "react-native";
+
+export default function Layout() {
+    return (
+        <Slot />
+    )
+}
