+++ src/app/addBook.js
@@ -1,0 +1,5 @@
+import { AddBook } from "../components/AddBook"
+
+export default function addBook() {
+    return <AddBook />;
+}
