+++ src/app/index.js
@@ -1,0 +1,5 @@
+import { Main } from "../components/Main";
+
+export default function Index() {
+    return <Main />;
+}
