+++ src/resources/SPIcon.png
